Move CudaReproTests to its own file
[ghstack-poisoned]
