+++ test/inductor/test_torchinductor.py
@@ -24,7 +24,6 @@ import torch
 import torch._dynamo
 import torch.nn as nn
 from torch._dispatch.python import enable_python_dispatcher
-from torch._dynamo.debug_utils import same_two_models
 from torch._dynamo.testing import rand_strided, same
 from torch._inductor.codegen.cpp import CppVecKernelChecker
 from torch._inductor.graph import GraphLowering
@@ -6941,8 +6940,6 @@ if HAS_CPU and not torch.backends.mps.is_available():
 
 
 if HAS_CUDA and not TEST_WITH_ASAN:
-    import triton
-    import triton.language as tl
 
     class SweepInputsCudaTest(SweepInputs2, TestCase):
         gen = InputGen(10, "cuda")
@@ -6953,479 +6950,7 @@ if HAS_CUDA and not TEST_WITH_ASAN:
         common = check_model_cuda
         device = "cuda"
 
-        def test_simplify_dims(self):
-            def fn(a):
-                return (a + 1,)
-
-            self.common(
-                fn, (torch.randn(2, 3, 10, 5, 6, device="cuda")[:, :, 2::2, :, :],)
-            )
-
-        @config.patch(permute_fusion=True)
-        def test_permute_fusion(self):
-            class Repro(torch.nn.Module):
-                def forward(self, view, reshape_2):
-                    permute = view.permute(0, 2, 1)
-                    view = None
-                    reshape = torch.reshape(permute, (-1, 642))
-                    bmm = torch.bmm(permute, reshape_2)
-                    return (bmm,)
-
-            args = [
-                ((1024, 642, 160), (102720, 160, 1), torch.float32, "cuda", True),
-                ((1024, 642, 20), (12840, 20, 1), torch.float32, "cuda", True),
-            ]
-            args = [
-                rand_strided(sh, st, dt, dev).requires_grad_(rg)
-                for (sh, st, dt, dev, rg) in args
-            ]
-
-            mod = Repro()
-            opt_mod = torch._dynamo.optimize("inductor")(mod)
-
-            ref = mod(*args)
-            res = opt_mod(*args)
-            self.assertTrue(same(ref, res))
-
-        @config.patch({"triton.autotune_pointwise": True})
-        def test_inplace_add_alpha_autotune(self):
-            def fn(x, y):
-                aten.add_.Tensor(x, y, alpha=0.55)
-                return (x,)
-
-            x1 = torch.zeros(2, 3, 4, 10, device="cuda")
-            x2 = torch.zeros(2, 3, 4, 10, device="cuda")
-            x3 = torch.zeros(2, 3, 4, 10, device="cuda")
-            y = torch.randn(2, 3, 4, 10, device="cuda").to(
-                memory_format=torch.channels_last
-            )
-            fn_fx = make_fx(fn)(x1, y)
-            fn_compiled = compile_fx_inner(fn_fx, [x1, y])
-            fn(x2, y)
-            fn_compiled([x3, y])
-            assert same(x2, x3)
-
-        @config.patch({"triton.autotune_pointwise": True})
-        def test_inplace_buffer_autotune(self):
-            def foo(x, y, z):
-                a = x @ y
-                return a.unsqueeze(0).unsqueeze(0) + z
-
-            x = torch.zeros(5, 5, device="cuda")
-            y = torch.zeros(5, 5, device="cuda")
-            z = torch.zeros(1, 1, 5, 5, device="cuda").to(
-                memory_format=torch.channels_last
-            )
-            self.common(
-                foo,
-                (x, y, z),
-                check_lowp=False,
-            )
-
-        def test_memory_history_inductor(self):
-            def called_inside_compile(x, w, b):
-                a = x @ w + b
-                return torch.sigmoid(a)
-
-            @torch.compile
-            def fn(x, w, b):
-                x = called_inside_compile(x, w, b)
-                return called_inside_compile(x, w, b)
-
-            w = torch.rand(3, 3, device="cuda")
-            b = torch.rand(3, device="cuda")
-            x = torch.rand(3, device="cuda")
-            try:
-                torch.cuda.memory.empty_cache()
-                torch.cuda.memory._record_memory_history(True)
-                r = fn(x, w, b)
-            finally:
-                torch.cuda.memory._record_memory_history(False)
-            snapshot = str(torch.cuda.memory._snapshot())
-            self.assertTrue("called_inside_compile" in snapshot)
-
     copy_tests(CommonTemplate, CudaTests, "cuda")
-
-    class CudaReproTests(TestCase):
-        common = check_model_cuda
-
-        def test_index_put_issue(self):
-            def forward(
-                self,
-                arg76_1,
-                expand_default,
-                full_like_default,
-                _to_copy_default_67,
-                zeros,
-            ):
-                sum_sym_int_19 = torch.ops.aten.sum(_to_copy_default_67, [0], True)
-                view_default_57 = torch.ops.aten.view.default(
-                    sum_sym_int_19, [512, 768]
-                )
-                where_self = torch.ops.aten.where.self(
-                    expand_default, view_default_57, full_like_default
-                )
-                clone_default_12 = torch.ops.aten.clone.default(zeros)
-                index_put__default = torch.ops.aten.index_put_.default(
-                    clone_default_12, [arg76_1], where_self, True
-                )
-                return (index_put__default,)
-
-            inps = [
-                (torch.Size([512]), torch.int64),
-                (torch.Size([512, 768]), torch.bool),
-                (torch.Size([512, 768]), torch.float16),
-                (torch.Size([4, 512, 768]), torch.float16),
-                (torch.Size([512, 768]), torch.float16),
-            ]
-            inps = [torch.zeros(())] + [
-                torch.ones(shape, dtype=dtype, device="cuda") for (shape, dtype) in inps
-            ]
-            mod = make_fx(forward)(*inps)
-            compiled = compile_fx_inner(mod, inps)
-            compiled(inps)
-
-        @requires_cuda()
-        def test_input_channels_last(self):
-            m = torch.nn.Sequential(
-                torch.nn.Conv2d(3, 3, 1, 1),
-                ToTuple(),
-            ).cuda()
-            inp = (
-                torch.randn([2, 3, 16, 16]).to(memory_format=torch.channels_last).cuda()
-            )
-
-            self.common(
-                m,
-                (inp,),
-                check_lowp=False,
-            )
-
-            @torch._dynamo.optimize()
-            def foo(m, inp):
-                return m(inp)
-
-            self.assertTrue(
-                foo(m, inp)[0].is_contiguous(memory_format=torch.channels_last)
-            )
-
-        # https://github.com/pytorch/torchdynamo/issues/1681#issuecomment-1283433527
-        @requires_cuda()
-        def test_unspec_inputs_interop(self):
-            class Repro(torch.nn.Module):
-                def forward(self, x, y):
-                    unsqueeze = torch.ops.aten.unsqueeze.default(x, 4)
-                    permute = torch.ops.aten.permute.default(unsqueeze, [0, 1, 2, 4, 3])
-                    add = torch.ops.aten.add.Tensor(y, 1)
-                    return [permute, add]
-
-            inps = [
-                rand_strided(
-                    (12, 3, 512, 64), (64, 196608, 768, 1), torch.float32, "cuda"
-                ),
-                rand_strided((), (), torch.int64, "cpu"),
-            ]
-            mod = make_fx(Repro().to(device="cuda"))(*inps)
-            compiled = compile_fx_inner(mod, inps)
-            compiled(inps)
-
-        @requires_cuda()
-        def test_backward_context(self):
-            def fn(x):
-                return x * 3
-
-            x = torch.randn(4, device="cuda", requires_grad=True)
-            gO = torch.rand_like(x)
-            opt_fn = torch.compile(fn)
-            out = opt_fn(x)
-            out.backward(gO)
-
-        @config.patch(fallback_random=True)
-        def test_dtype_factory_issue(self):
-            def forward():
-                randn = torch.ops.aten.randn.default(
-                    [12, 64, 1, 64],
-                    dtype=torch.float32,
-                    device=torch.device(type="cuda", index=0),
-                    pin_memory=False,
-                )
-                unsqueeze_default_2 = torch.ops.aten.unsqueeze.default(randn, -1)
-                return (unsqueeze_default_2,)
-
-            mod = make_fx(forward)()
-            compiled = compile_fx_inner(mod, ())
-            assert compiled([])[0].device.type == "cuda"
-
-        @config.patch({"triton.cudagraphs": True})
-        def test_no_device_idx_repro_cudagraphs(self):
-            class Repro(torch.nn.Module):
-                def __init__(self):
-                    super().__init__()
-
-                def forward(self):
-                    full = torch.ops.aten.full.default(
-                        [8, 512],
-                        1,
-                        dtype=torch.float32,
-                        layout=torch.strided,
-                        device=torch.device(type="cuda", index=0),
-                        pin_memory=False,
-                    )
-                    full_1 = torch.ops.aten.full.default(
-                        [8, 512],
-                        0,
-                        dtype=torch.int64,
-                        layout=torch.strided,
-                        device=torch.device(type="cuda", index=0),
-                        pin_memory=False,
-                    )
-                    return (full_1, full)
-
-            self.common(Repro(), ())
-
-        @config.patch({"triton.cudagraphs": True})
-        def test_expanded_inputs_cudagraphs(self):
-            @torch._dynamo.optimize("inductor")
-            def fn(x, y):
-                return x + y
-
-            inputs = (
-                rand_strided((5, 5, 5, 5), (0, 5, 0, 1), device="cuda"),
-                rand_strided((5, 5, 5, 5), (0, 5, 0, 1), device="cuda"),
-            )
-            self.assertTrue(same(fn(*inputs), inputs[0] + inputs[1]))
-
-        # TODO: Abstract this out, test more extensively
-        @torch._dynamo.config.patch(dynamic_shapes=True)
-        def test_dynamic_shapes(self):
-            torch._dynamo.reset()  # Needed since everywhere else uses "inductor"
-
-            def f(x):
-                return x.cos().view(x.shape).sin()
-
-            cnts = torch._dynamo.testing.CompileCounterWithBackend("inductor")
-
-            f2 = torch._dynamo.optimize(cnts)(f)
-
-            f2(torch.randn(32))
-
-            inp = torch.randn(16)
-            real_out = f(inp)
-            compiled_out = f2(inp)
-
-            self.assertEqual(cnts.frame_count, 1)
-            self.assertEqual(real_out, compiled_out)
-            torch._dynamo.reset()
-
-        @config.patch({"triton.cudagraphs": True, "size_asserts": False})
-        def test_expanded_inputs_cudagraphs_no_size_asserts(self):
-            @torch._dynamo.optimize("inductor")
-            def fn(x, y):
-                return x + y
-
-            inputs = (
-                rand_strided((5, 5, 5, 5), (0, 5, 0, 1), device="cuda"),
-                rand_strided((5, 5, 5, 5), (0, 5, 0, 1), device="cuda"),
-            )
-            self.assertTrue(same(fn(*inputs), inputs[0] + inputs[1]))
-
-        @config.patch({"triton.cudagraphs": True})
-        def test_inplace_updates_cudagraphs(self):
-            class Repro(torch.nn.Module):
-                def __init__(self):
-                    super().__init__()
-                    self.weight1 = torch.nn.Parameter(
-                        torch.randn(10, 20, requires_grad=True)
-                    )
-
-                def forward(self, x):
-                    x = torch.matmul(x, self.weight1)
-                    return x
-
-            from copy import deepcopy
-
-            model = Repro().cuda()
-            model_ref = deepcopy(model)
-            model_opt = torch._dynamo.optimize("inductor")(model)
-
-            input = torch.randn(10, 10, device="cuda", requires_grad=True)
-
-            for i in range(2):
-                output_ref = model_ref(input)
-                output_res = model_opt(input)
-                output_ref.sum().backward()
-                output_res.sum().backward()
-                for p_ref, p_res in zip(model_ref.parameters(), model_opt.parameters()):
-                    self.assertEqual(p_ref.grad, p_res.grad)
-                with torch.no_grad():
-                    for param in model_ref.parameters():
-                        param.add_(1.0)
-                    for param in model_opt.parameters():
-                        param.add_(1.0)
-
-        # https://github.com/pytorch/torchdynamo/issues/1850
-        def test_inductor_output_aliases_intermediate(self):
-            def foo(x):
-                out = x + x
-                return out.t()
-
-            foo_opt = torch._dynamo.optimize("inductor")(foo)
-
-            inpt = torch.randn(10, 10, device="cuda", requires_grad=True)
-            # TODO: this is broken, fix later
-            # out = foo_opt(inpt)
-            # out.add_(2)
-
-            out_ref = foo(inpt)
-            out_ref.add_(2)
-            # self.assertEqual(out_ref, out)
-
-        def test_accuracy_issue1(self):
-            class Repro(torch.nn.Module):
-                def __init__(self):
-                    super().__init__()
-                    self.linear = torch.nn.Linear(
-                        in_features=768, out_features=2, bias=True
-                    )
-
-                def forward(self, start_positions: torch.Tensor, x: torch.Tensor):
-                    linear = self.linear(x)
-                    split = linear.split(1, dim=-1)
-                    getitem = split[0]
-                    squeeze = getitem.squeeze(-1)
-                    clamp = start_positions.clamp(0, 128)
-                    cross_entropy = torch.nn.functional.cross_entropy(
-                        squeeze, clamp, None, None, 128, None, "mean", 0.0
-                    )
-                    return cross_entropy
-
-            mod = Repro().cuda()
-            opt_mod = torch._dynamo.optimize("inductor")(mod)
-            mod.eval()
-            opt_mod.eval()
-
-            args = [
-                ((1,), (1,), torch.int64, "cuda", False),
-                ((1, 128, 768), (98304, 768, 1), torch.float32, "cuda", True),
-            ]
-            args = [
-                rand_strided(sh, st, dt, dev).requires_grad_(rg)
-                for (sh, st, dt, dev, rg) in args
-            ]
-            with torch.cuda.amp.autocast(enabled=False):
-                assert same_two_models(mod, opt_mod, args), "Dynamo failed"
-
-        def test_autotune_inplace_kernel(self):
-            """
-            This UT tests autotune on an inplace kernel. The autotune should not contaminate
-            the input buffers when tuning with multiple configs. For more details, refer to
-            https://github.com/openai/triton/issues/781
-            https://github.com/pytorch/torchdynamo/issues/1670
-            """
-            from torch._C import _cuda_getCurrentRawStream as get_cuda_stream
-            from torch._inductor.triton_heuristics import CachingAutotuner, grid
-            from torch._inductor.utils import instance_descriptor
-
-            def autotune(configs, meta):
-                def decorator(fn):
-                    return CachingAutotuner(
-                        # force autotune by setting save_cache_hook to False
-                        fn,
-                        meta=meta,
-                        configs=configs,
-                        save_cache_hook=False,
-                        mutated_arg_names=["in_out_ptr0"],
-                    )
-
-                return decorator
-
-            @autotune(
-                configs=[
-                    triton.Config({"XBLOCK": 1}),
-                    triton.Config({"XBLOCK": 2}),
-                ],
-                meta={
-                    "signature": {0: "*fp32", 1: "*fp32", 2: "i32"},
-                    "device": 0,
-                    "configs": [
-                        instance_descriptor(divisible_by_16=(0, 1), equal_to_1=())
-                    ],
-                    "constants": {},
-                },
-            )
-            @triton.jit
-            def kernel(in_out_ptr0, in_ptr0, xnumel, XBLOCK: tl.constexpr):
-                pid = tl.program_id(0)
-                block_start = pid * XBLOCK
-                offsets = block_start + tl.arange(0, XBLOCK)
-                mask = offsets < xnumel
-                x = tl.load(in_out_ptr0 + offsets, mask=mask)
-                y = tl.load(in_ptr0 + offsets, mask=mask)
-                output = x + y
-                tl.store(in_out_ptr0 + offsets, output, mask=mask)
-
-            xnumel = 384
-            in0 = rand_strided((xnumel,), (1,), device="cuda", dtype=torch.float32)
-            inout1 = rand_strided((xnumel,), (1,), device="cuda", dtype=torch.float32)
-            inout2 = inout1.clone()
-
-            stream0 = get_cuda_stream(0)
-            kernel.run(inout1, in0, xnumel, grid=grid(xnumel), stream=stream0)
-            kernel.run(inout2, in0, xnumel, grid=grid(xnumel), stream=stream0)
-
-            assert same(
-                inout1, inout2, tol=0.001, equal_nan=True
-            ), "failed autotune with inplace kernel"
-
-        @requires_cuda()
-        def test_sort_stride_issue(self):
-            # This minified testcase comes from detectron2_maskrcnn_r_50_fpn
-            # There was a false error from our size_assert code
-            @torch._dynamo.optimize(nopython=True)
-            def forward(pred_objectness_logits_3_: torch.Tensor):
-                sort_3 = pred_objectness_logits_3_.sort(descending=True, dim=1)
-                getitem_12 = sort_3[0]
-                return getitem_12
-
-            args = [((1, 100), (0, 1), torch.float16, "cuda", False)]
-            args = [
-                rand_strided(sh, st, dt, dev).requires_grad_(rg)
-                for (sh, st, dt, dev, rg) in args
-            ]
-            result = forward(*args)
-            assert same(result, torch.sort(args[0], descending=True, dim=1)[0])
-
-        @requires_cuda()
-        def test_scalar_triton_index(self):
-            # The indirect indexing via a scalar like below used to lead to
-            # bad triton code that made triton segfault when compiling.
-            # See https://github.com/pytorch/torchdynamo/issues/1515
-            def fn(a):
-                zero = torch.zeros((16,), device=a.device, dtype=torch.int64)
-                return (a[zero],)
-
-            a = torch.randn((8,), dtype=torch.float32, device="cuda")
-
-            fn_optimized = torch._dynamo.optimize("inductor")(fn)
-            assert same(fn(a), fn_optimized(a))
-
-        @requires_cuda()
-        def test_indirect_indexing_dense_mask(self):
-            def fn(x, y):
-                ne = torch.ops.aten.ne.Scalar(x, 1)
-                sum_1 = torch.ops.aten.sum.dim_IntList(ne, [1])
-                sub = torch.ops.aten.sub.Tensor(sum_1, 1)
-                unsqueeze = torch.ops.aten.unsqueeze.default(sub, -1)
-                gather = torch.ops.aten.gather.default(x, 1, unsqueeze)
-                squeeze = torch.ops.aten.squeeze.default(gather)
-                out = torch.ops.aten.multiply(y, squeeze)
-                return (out,)
-
-            a = torch.zeros((1, 128), dtype=torch.int64, device="cuda")
-            b = torch.zeros((1, 128), dtype=torch.int64, device="cuda")
-
-            fn_optimized = torch._dynamo.optimize("inductor")(fn)
-            assert same(fn(a, b), fn_optimized(a, b))
 
     class TritonCodeGenTests(TestCase):
         from torch._inductor.triton_heuristics import CachingAutotuner
